EVEN MORE API ZOMG OMG LOL KKTHXBAI!
EVEN MORE API ZOMG OMG LOL KKTHXBAI!

+++ WebAPI/Models/ReservationRepository.cs
@@ -8,24 +8,25 @@ using System.Web;
 using System.Web.Http;
 using System.Web.Http.Results;
 using Models;
+using ModelSQL;
+using WebAPI.Security;
 
 namespace WebAPI.Models {
     public class ReservationRepository : IReservationRespository {
         private SqlConnection sqlConnection;
-        private SqlCommand sqlCommand;
-        private SqlDataReader reader;
-        private SqlParameter sqlParameter;
+        private readonly IAuthorizationSystem AuthorizationSystem;
 
         public ReservationRepository() {
             sqlConnection = Sql.GetInstance().GetConnection();
+            AuthorizationSystem = new AuthorizationSystem();
         }
 
         public IEnumerable<Reservation> GetAllReservations() {
-            sqlCommand = new SqlCommand("SELECT * FROM Reservations", sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("SELECT * FROM Reservations", sqlConnection);
             List<Reservation> reservations = null;
             try {
                 reservations = new List<Reservation>();
-                reader = sqlCommand.ExecuteReader();
+                SqlDataReader reader = sqlCommand.ExecuteReader();
                 while(reader.Read()) {
                     Reservation r = new Reservation();
                     r.Id = Convert.ToInt32(reader[0]);
@@ -42,11 +43,11 @@ namespace WebAPI.Models {
         }
 
         public IEnumerable<Reservation> GetBusReservation(String regNo, DateTime date) {
-            sqlCommand = new SqlCommand("SELECT * FROM Reservations r WHERE bus='AB12345' AND CONVERT(char(10),r.fromDate,126)=@date", sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("SELECT * FROM Reservations r WHERE bus='AB12345' AND CONVERT(char(10),r.fromDate,126)=@date", sqlConnection);
             List<Reservation> list = new List<Reservation>();
 
             // regno
-            sqlParameter = new SqlParameter("@regNo", SqlDbType.NVarChar);
+            SqlParameter sqlParameter = new SqlParameter("@regNo", SqlDbType.NVarChar);
             sqlParameter.Value = regNo;
             sqlCommand.Parameters.Add(sqlParameter);
 
@@ -54,7 +55,7 @@ namespace WebAPI.Models {
             sqlParameter = new SqlParameter("@date", SqlDbType.DateTime);
             sqlParameter.Value = date;
             sqlCommand.Parameters.Add(sqlParameter);
-
+            SqlDataReader reader = null;
             try {
                 reader = sqlCommand.ExecuteReader();
                 while(reader.Read()) {
@@ -69,26 +70,21 @@ namespace WebAPI.Models {
             } catch(Exception e) {
                 Debug.WriteLine(e.Message);
             } finally {
+                if(reader != null)
                 reader.Close();
-                string query = sqlCommand.CommandText;
-
-                foreach(SqlParameter p in sqlCommand.Parameters) {
-                    query = query.Replace(p.ParameterName, p.Value.ToString());
-                }
-                Debug.WriteLine(query);
             }
             return list;
         }
 
         public IEnumerable<Reservation> GetBusReservation(string regNo) {
-            sqlCommand = new SqlCommand("SELECT * FROM Reservations r WHERE bus='AB12345';", sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("SELECT * FROM Reservations r WHERE bus='AB12345';", sqlConnection);
             List<Reservation> list = new List<Reservation>();
 
             // regno
-            sqlParameter = new SqlParameter("@regNo", SqlDbType.NVarChar);
+            SqlParameter sqlParameter = new SqlParameter("@regNo", SqlDbType.NVarChar);
             sqlParameter.Value = regNo;
             sqlCommand.Parameters.Add(sqlParameter);
-
+            SqlDataReader reader = null;
             try {
                 reader = sqlCommand.ExecuteReader();
                 while(reader.Read()) {
@@ -103,14 +99,14 @@ namespace WebAPI.Models {
             } catch(Exception e) {
                 Debug.WriteLine(e.Message);
             } finally {
+                if(reader != null)
                 reader.Close();
             }
             return list;
         }
 
-
         private Bus GetBus(String regNo) {
-            sqlCommand = new SqlCommand("SELECT * FROM Busses WHERE regNo='" + regNo + "'", sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("SELECT * FROM Busses WHERE regNo='" + regNo + "'", sqlConnection);
             SqlDataReader reader = null;
             Bus b = null;
             try {
@@ -131,7 +127,7 @@ namespace WebAPI.Models {
         }
 
         private User GetUser(int id) {
-            sqlCommand = new SqlCommand("SELECT * FROM Users WHERE Mobile=" + id, sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("SELECT * FROM Users WHERE Mobile=" + id, sqlConnection);
             SqlDataReader reader = null;
             User u = null;
             try {
@@ -151,27 +147,49 @@ namespace WebAPI.Models {
                     reader.Close();
                 }
             }
-
             return u;
         }
 
-        public bool DeleteReservation(int reservationId) {
-            sqlCommand = new SqlCommand("DELETE FROM Reservations WHERE id=" + reservationId, sqlConnection);
-            int i = -1;
+        public int DeleteReservation(int reservationId) {
+            // check if it's the users own reservation
+            SqlCommand sqlCommand = new SqlCommand("SELECT username FROM reservations WHERE id=" + reservationId + ";",sqlConnection);
+            string username = HttpContext.Current.User.Identity.Name;
+            bool ownReservation = false;
+            SqlDataReader reader = null;
             try {
-                i = sqlCommand.ExecuteNonQuery();
-            } catch(Exception e) {
-                Debug.WriteLine(e.Message);
+                reader = sqlCommand.ExecuteReader();
+                if(reader.Read()) {
+                    ownReservation = Convert.ToInt32(reader[0]) ==
+                                     Convert.ToInt32(username);
+                }
+            } catch(SqlException e) {
+                Debug.WriteLine("Failed to check username: " + e.Message);
+            } finally {
+                if(reader != null)
+                reader.Close();
             }
-            return i > 0;
+
+            // check logic - here we check if you are admin or it's your own reservation
+            if(!ownReservation && !AuthorizationSystem.IsAdmin(username)) {
+                return 0;
+            } else {
+                sqlCommand = new SqlCommand("DELETE FROM Reservations WHERE id=" + reservationId, sqlConnection);
+                int i = -2;
+                try {
+                    i = sqlCommand.ExecuteNonQuery();
+                } catch(Exception e) {
+                    Debug.WriteLine(e.Message);
+                }
+                return i;
+            }
         }
 
         public Reservation PostReservation(Reservation reservation) {
 
-            sqlCommand = new SqlCommand("INSERT INTO Reservations VALUES(@username,@bus,@fromDate,@toDate); SELECT Scope_Identity();", sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("INSERT INTO Reservations VALUES(@username,@bus,@fromDate,@toDate); SELECT Scope_Identity();", sqlConnection);
 
             // Username
-            sqlParameter = new SqlParameter("@username", SqlDbType.Int);
+            SqlParameter sqlParameter = new SqlParameter("@username", SqlDbType.Int);
             sqlParameter.Value = reservation.User.Mobile;
             sqlCommand.Parameters.Add(sqlParameter);
 
@@ -193,8 +211,6 @@ namespace WebAPI.Models {
             int id = -1;
             try {
                 id = Convert.ToInt32(sqlCommand.ExecuteScalar());
-
-                Debug.WriteLine(id);
             } catch(Exception e) {
                 Debug.WriteLine(e.Message);
             }
@@ -202,11 +218,10 @@ namespace WebAPI.Models {
         }
 
         public Reservation PostReservation(ReservationApiClass reservation) {
-
-            sqlCommand = new SqlCommand("INSERT INTO Reservations VALUES(@username,@bus,@fromDate,@toDate); SELECT Scope_Identity();", sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("INSERT INTO Reservations VALUES(@username,@bus,@fromDate,@toDate); SELECT Scope_Identity();", sqlConnection);
 
             // Username
-            sqlParameter = new SqlParameter("@username", SqlDbType.Int);
+            SqlParameter sqlParameter = new SqlParameter("@username", SqlDbType.Int);
             sqlParameter.Value = reservation.Mobile;
             sqlCommand.Parameters.Add(sqlParameter);
 
@@ -225,22 +240,20 @@ namespace WebAPI.Models {
             sqlParameter.Value = reservation.ToDate;
             sqlCommand.Parameters.Add(sqlParameter);
 
-            int id = -1;
+            Reservation res = null;
             try {
-                id = Convert.ToInt32(sqlCommand.ExecuteScalar());
-                
-                Debug.WriteLine(id);
-                return GetReservation(id);
+                int id = Convert.ToInt32(sqlCommand.ExecuteScalar());
+                res = GetReservation(id);
             } catch(Exception e) {
                 Debug.WriteLine(e.Message);
-                return null;
             }
+            return res;
         }
 
-
         public Reservation GetReservation(int id) {
-            sqlCommand = new SqlCommand("SELECT * FROM Reservations WHERE id=" + id, sqlConnection);
+            SqlCommand sqlCommand = new SqlCommand("SELECT * FROM Reservations WHERE id=" + id, sqlConnection);
             Reservation r = null;
+            SqlDataReader reader = null;
             try {
                 reader = sqlCommand.ExecuteReader();
                 while(reader.Read()) {
@@ -254,7 +267,8 @@ namespace WebAPI.Models {
             } catch(Exception e) {
                 Debug.WriteLine(e.Message);
             } finally {
-                reader.Close();
+                if(reader != null)
+                    reader.Close();
             }
             return r;
         }
